refactor: rename makeRequest file to config file
